Updating Enzyme's mount() so it "works" with React 16
Summary:

This no longer renders to DOM (yet). As a result, the component tree that mount()
returns is a bit odd, and contains a mix of elements and instances. Or, actually,
I'm not sure it is odd or it is just how Fiber works.

Simulate() in mount() is now totally unrealistic, just like in shallow().

Matching by prop no longer matches only DOM nodes (mostly because the behavior
didn't seem consistent to me, as sometimes our tests want to match only DOM
nodes, like when clicking, and sometimes they don't, like when reading props
or finding things within a specific item in a list). While simulate was also
changed to work with multiple nodes (if they are all nested within each other)
most other methods will complain and ask you to provide a more specific selector.

Also, the notion of "DOM nodes" is a bit undefined given we are no longer
rendering to 'jsdom', and within the Fiber hierarchy sometimes even for a
given spec (testID="foo", for example) there are either 1 or 2 components
that match it and are marked as "mounted". I don't fully understand why,
but usually only one of these 2 components has a 'onPress' prop we can fire
directly. Due to this mess that I quite don't understand (and I don't think
I can without digging deep into how Fiber is implemented and some of its
internals), the click-as-press hack has been backed into our version of
Enzyme, as this proved to provide the most consistent behavior overall.

Direct descendent selectors no longer work, simply because ReactTestUtils
provides no helper for this *and* it is really, really difficult to understand
what's going on when I read at the implementation of how ReactTestUtils
navigate within the tree of fibers to find components. This will throw a
NotImplementedError right now.

Test Plan:

All Exam tests pass with this "version" of enzyme